Correção exclusão de foto

+++ src/main/java/com/brewer/storage/s3/FotoStorageS3.java
@@ -79,7 +79,9 @@ public class FotoStorageS3 implements FotoStorage {
     @Override
     public void excluir(String foto) {
         logger.debug(foto);
-        amazonS3.deleteObjects(new DeleteObjectsRequest(bucketName).withKeys(foto, Constantes.THUMBNAIL_PREFIX + foto));
+        if (!StringUtils.isEmpty(foto)) {
+            amazonS3.deleteObjects(new DeleteObjectsRequest(bucketName).withKeys(foto, Constantes.THUMBNAIL_PREFIX + foto));
+        }
     }
 
     @Override
